requirements.txt
requests==2.28.1
python-dotenv==0.20.0
pytest==7.1.2
